+++ src/Repository/BucketRepository.php
@@ -1,0 +1,12 @@
+<?php
+
+
+namespace Coffreo\CephOdm\Repository;
+
+/**
+ * AbstractRepositoryWrapper for BucketRepository
+ */
+class BucketRepository extends AbstractRepositoryDecorator
+{
+
+}