+++ src/Repository/FileRepository.php
@@ -1,0 +1,12 @@
+<?php
+
+
+namespace Coffreo\CephOdm\Repository;
+
+/**
+ * AbstractRepositoryWrapper for FileRepository
+ */
+class FileRepository extends AbstractRepositoryDecorator
+{
+
+}